backend/requirements.txt
# Vast.ai GPU Worker - Python Dependencies
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
trimesh>=4.0.0
pyglet>=2.0.0
requests>=2.31.0
